requirements.txt
alembic==1.14.0
APScheduler==3.11.0
bcrypt==4.2.1
blinker==1.9.0
certifi==2024.12.14
charset-normalizer==2.0.12
click==8.1.8
dnspython==2.7.0
dominate==2.9.1
email_validator==2.2.0
Flask==3.0.3
Flask-APScheduler==1.13.1
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-Moment==1.0.2
Flask-PyMongo==2.3.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.1.1
gunicorn==20.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
psycopg2-binary==2.9.9
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==0.19.0
requests==2.26.0
six==1.17.0
SQLAlchemy==2.0.32
stripe==2.61.0
typing_extensions==4.12.2
tzlocal==5.2
urllib3==1.26.20
visitor==0.1.3
Werkzeug==3.1.3
WTForms==3.2.1
